Move package as is
